Добавляет разбивку на модули
